+++ modules/core/src/smithy4s/Document.scala
@@ -63,6 +63,11 @@ sealed trait Document extends Product with Serializable {
       value.map { case (k, v) => k + "=" + v.show }.mkString("{", ", ", "}")
   }
 
+  /**
+    * Nests this document under the given key.
+    * The result is a single-entry Document object.
+    */
+  def nest(k: String): Document = obj(k -> this)
 }
 
 object Document {
@@ -105,7 +110,7 @@ object Document {
   trait EncoderCompiler extends CachedSchemaCompiler[Encoder] {
     @deprecated(
       message = """Use withFieldFilter instead.
-      
+
   Mapping:
    - explicitDefaultsEncoding = false -> FieldFilter.Default
    - explicitDefaultsEncoding = true -> FieldFilter.EncodeAll