+++ modules/json/src/smithy4s/json/internals/SchemaVisitorJCodec.scala
@@ -982,43 +982,61 @@ private[smithy4s] class SchemaVisitorJCodec(
 
   private type Writer[A] = A => JsonWriter => Unit
 
-  private abstract class TaggedUnionJCodec[U](alternatives: Vector[Alt[U, _]])(
+  private abstract class UnionJCodec[U](alternatives: Vector[Alt[U, _]])(
       dispatch: Alt.Dispatcher[U]
   ) extends JCodec[U] {
 
-    val expecting = "tagged-union"
-
-    override def canBeKey: Boolean = false
-
-    def jsonLabel[A](alt: Alt[U, A]): String =
+    private def jsonLabel[A](alt: Alt[U, A]): String =
       alt.hints.get(JsonName) match {
         case None    => alt.label
         case Some(x) => x.value
       }
 
+    private def handler[A](alt: Alt[U, A]) = {
+      val codec = apply(alt.schema)
+      (cursor: Cursor, reader: JsonReader) =>
+        alt.inject(cursor.decode(codec, reader))
+    }
+
+    protected val unknownTagHandler =
+      alternatives
+        .find(_.hints.has(JsonUnknown))
+        .map(handler(_))
+        .orNull
+
     protected val handlerMap =
       new util.HashMap[String, (Cursor, JsonReader) => U] {
-        def handler[A](alt: Alt[U, A]) = {
-          val codec = apply(alt.schema)
-          (cursor: Cursor, reader: JsonReader) =>
-            alt.inject(cursor.decode(codec, reader))
-        }
-
-        alternatives.foreach(alt => put(jsonLabel(alt), handler(alt)))
+        alternatives
+          .filterNot(_.hints.has(JsonUnknown))
+          .foreach(alt => put(jsonLabel(alt), handler(alt)))
       }
+
+  }
+
+  private abstract class TaggedUnionJCodec[U](alternatives: Vector[Alt[U, _]])(
+      dispatch: Alt.Dispatcher[U]
+  ) extends UnionJCodec[U](alternatives)(dispatch) {
+
+    val expecting = "tagged-union"
+
+    override def canBeKey: Boolean = false
 
     protected val precompiler = new smithy4s.schema.Alt.Precompiler[Writer] {
       def apply[A](label: String, instance: Schema[A]): Writer[A] = {
-        val jsonLabel =
-          instance.hints.get(JsonName).map(_.value).getOrElse(label)
         val jcodecA = instance.compile(self)
-        a =>
-          out => {
-            out.writeObjectStart()
-            out.writeKey(jsonLabel)
-            jcodecA.encodeValue(a, out)
-            out.writeObjectEnd()
-          }
+
+        if (!instance.hints.has(JsonUnknown)) {
+          val key = instance.hints.get(JsonName).map(_.value).getOrElse(label)
+          a =>
+            out => {
+              out.writeObjectStart()
+              out.writeKey(key)
+              jcodecA.encodeValue(a, out)
+              out.writeObjectEnd()
+            }
+        } else { a => out =>
+          jcodecA.encodeValue(a, out)
+        }
       }
     }
     protected val writer = dispatch.compile(precompiler)
@@ -1040,27 +1058,42 @@ private[smithy4s] class SchemaVisitorJCodec(
   )(dispatch: Alt.Dispatcher[U]): JCodec[U] =
     new TaggedUnionJCodec[U](alternatives)(dispatch) {
 
-      def decodeValue(cursor: Cursor, in: JsonReader): U =
+      def decodeValue(cursor: Cursor, in: JsonReader): U = {
+        in.setMark()
         if (in.isNextToken('{')) {
           if (in.isNextToken('}'))
             in.decodeError("Expected a single key/value pair")
           else {
             in.rollbackToken()
             val key = in.readKeyAsString()
-            cursor.push(key)
             val handler = handlerMap.get(key)
-            if (handler eq null) in.discriminatorValueError(key)
-            val result = handler(cursor, in)
-            cursor.pop()
-            if (in.isNextToken('}')) result
+
+            // happy path
+            if (handler ne null) {
+              cursor.push(key)
+              val result = handler(cursor, in)
+              cursor.pop()
+              if (in.isNextToken('}')) result
+              else {
+                in.rollbackToken()
+                in.decodeError(s"Expected no other field after $key")
+              }
+            }
+            // @jsonUnknown path
+            else if (unknownTagHandler ne null) {
+              in.rollbackToMark()
+              unknownTagHandler(cursor, in)
+            }
+            // neither a known tag nor jsonUnknown
             else {
-              in.rollbackToken()
-              in.decodeError(s"Expected no other field after $key")
+              in.discriminatorValueError(key)
             }
           }
         } else in.decodeError("Expected JSON object")
+      }
     }
 
+  // todo: open unions here too
   private def lenientTaggedUnion[U](
       alternatives: Vector[Alt[U, _]]
   )(dispatch: Alt.Dispatcher[U]): JCodec[U] =
@@ -1161,29 +1194,10 @@ private[smithy4s] class SchemaVisitorJCodec(
       alternatives: Vector[Alt[U, _]],
       discriminated: Discriminated
   )(dispatch: Alt.Dispatcher[U]): JCodec[U] =
-    new JCodec[U] {
+    new UnionJCodec[U](alternatives)(dispatch) {
       def expecting: String = "discriminated-union"
 
       override def canBeKey: Boolean = false
-
-      def jsonLabel[A](alt: Alt[U, A]): String =
-        alt.hints.get(JsonName) match {
-          case None    => alt.label
-          case Some(x) => x.value
-        }
-
-      private[this] val handlerMap =
-        new util.HashMap[String, (Cursor, JsonReader) => U] {
-          def handler[A](
-              alt: Alt[U, A]
-          ): (Cursor, JsonReader) => U = {
-            val codec = apply(alt.schema)
-            (cursor: Cursor, reader: JsonReader) =>
-              alt.inject(cursor.decode(codec, reader))
-          }
-
-          alternatives.foreach(alt => put(jsonLabel(alt), handler(alt)))
-        }
 
       def decodeValue(cursor: Cursor, in: JsonReader): U =
         if (in.isNextToken('{')) {
@@ -1193,7 +1207,8 @@ private[smithy4s] class SchemaVisitorJCodec(
             in.rollbackToMark()
             in.rollbackToken()
             cursor.push(key)
-            val handler = handlerMap.get(key)
+            var handler = handlerMap.get(key)
+            if (handler eq null) handler = unknownTagHandler
             if (handler eq null) in.discriminatorValueError(key)
             val result = handler(cursor, in)
             cursor.pop()